examples/voice_agents/requirements.txt
livekit-agents[openai, cartesia, elevenlabs, deepgram, silero, turn-detector, mcp, assemblyai]>=1.0
python-dotenv>=1.0
duckduckgo-search>=8.0